requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.20
librosa
numpy
python-multipart
pydub